data-ingestion/requirements.txt
# Data collection
yfinance==0.2.28  # Sử dụng phiên bản cũ hơn
pandas>=1.5.0
requests>=2.28.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0

# HDFS
hdfs>=2.7.0

# Configuration
python-dotenv>=0.21.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Scheduler
apscheduler>=3.10.0

# Utilities
python-dateutil>=2.8.0